split theme plugin
