+++ libs/platform/core/src/orchestration/components/components.ts
@@ -27,13 +27,13 @@ export class ComponentsPlugin extends ComponentsObserver implements AppPlugin {
     programmaticLoad: true,
   };
   protected theme?: ThemePlugin;
-  rootSelector = '';
+  protected rootSelector = '';
 
   constructor(
-    componentsInfo: ComponentsInfo,
-    public options: ComponentsOptions
+    protected componentsInfo: ComponentsInfo,
+    protected options: ComponentsOptions
   ) {
-    super();
+    super(options);
     this.registerComponents(componentsInfo);
   }
 
@@ -44,10 +44,7 @@ export class ComponentsPlugin extends ComponentsObserver implements AppPlugin {
   async apply(app: App): Promise<void> {
     this.theme = app.findPlugin(ThemePlugin);
 
-    this.rootSelector =
-      typeof this.options.root === 'string'
-        ? this.options.root
-        : this.processDef(this.options.root).name;
+    const root = this.getRoot();
 
     if (this.options.preload) {
       await this.preloadComponents();
@@ -55,11 +52,11 @@ export class ComponentsPlugin extends ComponentsObserver implements AppPlugin {
       return;
     }
 
-    const rootElement = document.querySelector?.(this.rootSelector);
+    const rootElement = document.querySelector?.(root);
 
     if (!rootElement) {
       throw new ComponentsPluginError(
-        `Cannot find root element by selector '${this.rootSelector}'!`
+        `Cannot find root element by selector '${root}'!`
       );
     }
 
@@ -73,6 +70,16 @@ export class ComponentsPlugin extends ComponentsObserver implements AppPlugin {
       const def = this.processDef(info);
       this.componentDefMap.set(def.name, def);
     });
+  }
+
+  getOptions(): ComponentsOptions {
+    return this.options;
+  }
+
+  getRoot(): string {
+    return typeof this.options.root === 'string'
+      ? this.options.root
+      : this.processDef(this.options.root).name;
   }
 
   async loadComponent(name: string): Promise<ComponentType | undefined> {
