+++ libs/template/themes/design-tokens/src/storefront/index.ts
@@ -1,47 +1,38 @@
 import { DesignToken } from '@spryker-oryx/core';
-import { Size } from '@spryker-oryx/utilities';
 import { darkColor } from './color-dark.tokens';
 import { color } from './color.tokens';
-import { layoutMdTokens, layoutSmTokens, layoutTokens } from './layout.tokens';
-import { darkTokens } from './other-dark';
-import { tokens } from './other.tokens';
-import {
-  typographyMediumTokens,
-  typographySmallTokens,
-  typographyTokens,
-} from './typography.tokens';
 
 export const storefrontTokens: DesignToken[] = [
   {
     color,
-    ...tokens,
-    ...typographyTokens,
+    // ...tokens,
+    // ...typographyTokens,
   },
   {
     media: {
       mode: 'dark',
     },
     color: darkColor,
-    ...darkTokens,
+    // ...darkTokens,
   },
-  {
-    media: {
-      screen: Size.Sm,
-    },
-    ...layoutSmTokens,
-    ...typographySmallTokens,
-  },
-  {
-    media: {
-      screen: Size.Md,
-    },
-    ...layoutMdTokens,
-    ...typographyMediumTokens,
-  },
-  {
-    media: {
-      screen: Size.Lg,
-    },
-    ...layoutTokens,
-  },
+  // {
+  //   media: {
+  //     screen: Size.Sm,
+  //   },
+  //   ...layoutSmTokens,
+  //   ...typographySmallTokens,
+  // },
+  // {
+  //   media: {
+  //     screen: Size.Md,
+  //   },
+  //   ...layoutMdTokens,
+  //   ...typographyMediumTokens,
+  // },
+  // {
+  //   media: {
+  //     screen: Size.Lg,
+  //   },
+  //   ...layoutTokens,
+  // },
 ];
